create graphics.py, move plot_bitinformation
