+++ bitinformation_pipeline/__init__.py
@@ -1,10 +1,6 @@
 """Top-level package for bitinformation_pipeline."""
 
-from .bitinformation_pipeline import (
-    _get_keepbits,
-    get_bitinformation,
-    get_keepbits,
-    plot_bitinformation,
-)
+from .bitinformation_pipeline import _get_keepbits, get_bitinformation, get_keepbits
 from .bitround import jl_bitround, xr_bitround
+from .graphics import plot_bitinformation
 from .save_compressed import get_compress_encoding
